Corrections on interval calculation
I made the ci_bayes function applicable to BVR, BSC and SMAC. Now we can use the same function for all bayesian methods.

Also, previously, we were giving the STANDARDIZED data to ci_bayes. Therefore, the credible intervals we were getting for ci_bayes were not for the original scale of the data. I corrected this.

This might have to be corrected for the frequentist methods as well.