fixing wrong loop
When we calculate upper and lower, we had a loop over time periods, but we do not need that. We were essentially updating the lower and upper CIs for all time periods using the SE from the last time period.

By dropping the for loop, the issue should be fixed.